+++ pkarr/src/client.rs
@@ -2,15 +2,15 @@
 
 pub mod cache;
 
-#[cfg(feature = "relay")]
+#[cfg(any(feature = "relays", target_arch = "wasm32"))]
 mod shared;
 
 #[cfg(not(target_arch = "wasm32"))]
 pub(crate) mod native;
 #[cfg(not(target_arch = "wasm32"))]
-pub use native::{Client, Config};
+pub use native::{Client, ClientBuilder, Config, Info};
 
 #[cfg(target_arch = "wasm32")]
 pub(crate) mod web;
 #[cfg(target_arch = "wasm32")]
-pub use web::{Client, Config};
+pub use web::{Client, ClientBuilder, Config};